pyproject.toml
[tool.poetry]
name = "drf-hw(coursestream)"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.2.1"
django = "^5.2.8"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
pillow = "^12.0.0"
faker = "^38.2.0"
django-filter = "^25.2"
djangorestframework-simplejwt = "^5.5.1"
ipython = "^9.8.0"
coverage = "^7.13.0"


[tool.poetry.group.lint.dependencies]
isort = "^7.0.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"
black = "^25.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
ignore_missing_imports = true

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
